*** empty log message ***
git-svn-id: https://crosswire.org/svn/sword/trunk@732 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ tests/complzss.cpp
@@ -23,8 +23,8 @@ class FileCompress: public LZSSCompress {
 public:
 	FileCompress(char *);
 	~FileCompress();
-	int GetChars(char *, int len);
-	int SendChars(char *, int len);
+	unsigned long GetChars(char *, unsigned long len);
+	unsigned long SendChars(char *, unsigned long len);
 	void Encode();
 	void Decode();
 };
@@ -52,13 +52,13 @@ FileCompress::~FileCompress()
 }
 
 
-int FileCompress::GetChars(char *buf, int len) 
+unsigned long FileCompress::GetChars(char *buf, unsigned long len) 
 {
 	return read(ifd, buf, len);
 }
 
 
-int FileCompress::SendChars(char *buf, int len) 
+unsigned long FileCompress::SendChars(char *buf, unsigned long len) 
 {
 	return write(ofd, buf, len);
 }
